tp_bot/requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
Babel==2.9.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
loguru==0.5.3
multidict==5.1.0
psycopg2==2.9.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.5
win32-setctime==1.0.3
yarl==1.6.3
